[HOTSPOT-181] fix: familyDataAmount int -> long으로 변경

+++ src/main/java/hotspot/user/policy/controller/response/FamilyAppliedPolicyResponse.java
@@ -12,7 +12,7 @@ import lombok.Builder;
 public record FamilyAppliedPolicyResponse(
         Long familyId,
         int familyNum,
-        int familyDataAmount,
+        double familyDataAmount,
         PriorityType priorityType,
         List<AppliedPolicyResponse> memberPolicies
 ) {
